implement naive vertex indexing. Memory consumption is reduced but load times have suffered.
